requirements.txt
streamlit
google-generativeai
Pillow
gspread
google-auth
pandas